BST.constructor
Finally, after much over engineering, BST.constructor is a go. It accepts a val and options. By default, it is a normal BST that does not allow for duplicates. However, options can be configured to allow duplicates; and measure can be adjusted to allow for the BST to store non-numerical values